requirements.txt
httplib2==0.12.0
SQLAlchemy==1.1.11
PyWavelets==0.5.2
obspy==1.1.0
scipy==0.19.1
spectrum==0.6.5
numpy==1.15.4
matplotlib==2.0.2
PyMySQL==0.9.2
tensorflow_gpu==1.12.0
pywt==1.0.6